Assert content encoded etag is different than unecoded etag
Note: Content codings are a property of the representation data, so a strong
entity-tag for a content-encoded representation has to be distinct from the
entity tag of an unencoded representation to prevent potential conflicts
during cache updates and range requests. In contrast, transfer codings apply
only during message transfer and do not result in distinct entity-tags.