Fix source_map_concat
Line numbers and columns are stored in an array so it is `[<line number>, <column>]`.

Source map line numbers can never be 0, so we should always start with 1. When we concat this source map:

```
[
  { source: 'a.js', generated: [1, 0], original: [1,  0] },
  { source: 'b.js', generated: [2, 0], original: [20, 0] },
  { source: 'c.js', generated: [3, 0], original: [30, 0] }
]
```

With this one:

```
{ source: 'd.js', generated: [1, 0], original: [1, 0] }
```

We would expect that since `d.js` generates onto the first line when compiled by itself, it would go onto the next free line when concatenated with the first free line above. So the result should be:

```
[
  { source: 'a.js', generated: [1, 0], original: [1,  0] },
  { source: 'b.js', generated: [2, 0], original: [20, 0] },
  { source: 'c.js', generated: [3, 0], original: [30, 0] },
  { source: 'd.js', generated: [4, 0], original: [1,  0] }
]
```

The original test was invalid since it used 0 as a starting line number. To make this test pass we don't want to increment the offset by + 1 otherwise we would have `d.js` be on the 5th line instead of the 4th line since it already effectively has an offset line number of 1 (remember 0 is impossible).

The modified hashes in the `test_asset.rb` file is due to the different contents of source maps. I didn't verify these hashes or the results of their source maps were correct, not sure of a good way to do that ¯\_(ツ)_/¯