Fix stubbing of Zlib::GzipWriter with expected args
The manifest test for raising exceptions from Zlib::GzipWriter was
catching an ArgumentException from Ruby when the stubbing lambda
(that accepted no arguments) was called with the normal two.  This fixes
the test to assert the correct exception and message was raised.

Observed on minitest 5.4.3 where assert_raises let the ArgumentError
bubble up and fail the test.