Explicit API to error when an asset is not found
The goal here is that we can eventually extend this to look for common problems. For example in the future if you try to load `asset_path(assets/javascripts/application.js)` you might get an error telling you that you that you should leave off the paths in front of the asset. We could also look for assets with the same name and different file extensions as well as do a `did you mean` type levenshtein lookup for possible asset mis-spellings.

We can't do any of that today because there is no guarantee that you expect `find_asset` to return a valid result. By introducing an explicit api `find_asset!` we can guarantee the user either wants an asset or an error.