[close #440] Fix circular directives
We de-couple the directive parsing and "resolving" from the "loading" so that a file does not have the ability to accidentally load itself when it is itself trying to load.

In the DirectiveProcessor we no longer run `load` instead we store the things we want to load on the asset metadata. Later we pull those things out and load them. Here's a tree representation of what's now happening

- env.load("a.js") (pipeline=default)
  - env.load("a.js") (pipeline=source)
    - No processors for pipeline=source
    - Return pipeline=source
  - Processors for pipeline=default: [Bundle]
    - env.load("a.js") (pipeline=self)
      - Processors for pipeline=self: [DirectiveProcessor]
        - //= link b.js
          - b.s is added to `to_load` and `to_link` sets
      - pipeline=self asset returns
    - URIs pulled out of `to_load`, `to_link` sets and loaded
    - URIs pulled out of `reuired` set and loaded (existing behavior)
    - Merge required files to form top level asset
- Return top level asset

There is one concern here in that the extra `to_link` and `to_load` metadata gets stored in the cache for the `pipeline=self` asset. Also this asset would have traditionally have full URIs stored in the `links` metadata key, which can't be done yet because it expects fully expanded uris with digests. However, none of the tests fail, so I guess we're good? 

I think this is okay because we never actually use the file cache of `pipeline=self` (maybe?) instead we're always using the top level `pipeline=source` cache (I think). If that's the case we could save some processing overhead by not storing the same asset in the cache three different times. I honestly see no reason why this has to be recursive. It just makes the coder harder to reason about...but that's for a different PR/day.