Remove source map hack
When trying to load a source map file such as `application.js.map` it will load the `application.js` file and run the `SourceMapProcessor` on it as a transformer, before it can do that it needs to load dependencies in `application.js` the accept type for the directive processor will be "application/js-sourcemap+json". If the `application.js file has dependencies such as `//= require project` then it will try to find a `project` file with the same content type of the parent file which is `application/js-sourcemap+json` which would mean it's looking for a `project.js.map` which does not exist on disk. Instead it actually needs the `project.js` file.

This change checks if source map content type is being used in the directive processor and if so, it converts it to the original content type of the file. So a `application/js-sourcemap+json` will be converted to `application/javascript` for the directive processor.