Allow weak tags to be processed by sprockets.
Since nginx 1.7.3, strong ETags are transformed into weak ETags "starting with W/" if the request is compressed. That means that when the browser re-requests the assets in development mode, sprockets will see an ETag starting with "W/..." which will never match, and as such, cache will never work. So I'm relaxing the regex here a little to handle that.