Add extra mimetypes.
* Adds `audio/mp4` -> `.m4a`
* Adds `audio/ogg` -> `.ogg`, `.oga`
* Adds `audio/flac` -> `.flac`
* Adds `audio/aac` ->` ,aac`

The upcoming release of Jekyll-Assets will support `<audio>` and `<video>`. As such, upon our testing we noticed that there are a few mime-types missing from the list that we would like to support for people. This adds support for extra mime-types, common mime-types we use, and would like to support without extra code inside of Jekyll-Assets, we find that we often encode in `aac`, and `m4a`, we also find that a lot of Linux users still like `ogg`, and personally when I do stuff, it's always in `flac`.