Fix Stable Sort Test
The test mentioned in https://github.com/rails/sprockets/pull/573/files#r213790852 is actually currently only testing stable sort, and not actually testing the mtime based deletion which it presumed that it was doing, this is because the mtime of the file is set based off of Time.now and not based off of the mtime of the file as it was written. We can account for this by using time cop to "travel" forward in time to change the mtime both for the file on disk and for the value stored in the manifest.