Fix asset integrity not compliant with Subresource integrity RFC:
Few years ago, in 64fadf908, a change was introduced and modified the
digest of an asset. The returned value was the result of computing the
asset's content AND the sprockets environment version.
This fix was made in order to properly bust the asset cache when
the sprockets environment version was modified.

As a result of this change, the `Asset#digest` which is used to look
for the asset's integrity doesn't comply with the RFC as the returned
value should be hash representing the asset's content only.

More detail in #555

Based on feedbacks, I have reverted the change that modifies the
`Asset#digest` to remain a pure function of the digest.
I however modified the `digest_path` so that the hash represents the
sprockets environment version AND the asset's content.

The main difference in the manifest is that the digest_path is no
longer the same as the digest

Before
```json
{"files":{"application-a2f9b346fff9b3e6ea8e29c613dd3b0abdec1207fb111ca10acfb249ac5ace1c.js":{digest":"a2f9b346fff9b3e6ea8e29c613dd3b0abdec1207fb111ca10acfb249ac5ace1c"}}
```

After
```json
{"files":{"application-dsadsad91984391rfasfasf87f8asfasfdasd81dsjqisdasds87dasdas.js":{digest":"a2f9b346fff9b3e6ea8e29c613dd3b0abdec1207fb111ca10acfb249ac5ace1c"}}
```

Co-authored-by: Yuri S. <fudoshiki.ari@gmail.com>